Uses reboot handler in grsecurity role
Notifies the "reboot" handler that is exposed via the common role, since
the grsecurity role is included in the same play. Note that this makes
the play logic brittle: moving the grsecurity role to a separate play
will cause breakage. Fortunately our config tests will catch that, and
down the road relying more heavily on upstream will obviate much of the
need for custom logic such as this.