Clarified wording, steps, and improved continuity
Made changes to remove a bit of terseness and clarified some wording
so that an admin with little experience with Tails and GNU/Linux can
follow along with better continuity of steps. Also, added text to
accompany images. This should allow an admin without access to images
the ability to follow along with less ambiguity.