Creates intelligent handler for bouncing tor
If we restart the tor service during a playbook run that's connected
over Tor, the playbook will immediately fail. Any changes to the torrc
will notify the "restart tor" handler, and currently the *only* time
that won't break is during the very first playbook run, at install time.

Let's make the handler notification smarter, and if we're connected over
Tor, we'll bounce the service async (via fire-and-forget), then wait 90
seconds until trying to connect again. Normally we could use
local_action/wait_for to wait _up to_ 300 seconds or so, and continue
execution if the port opens within a smaller time interval, e.g. 90
seconds; but unfortunately the wait_for module is not smart enough to
build the socket query based on extra connection info such as is
required for routing the connection over Tor. Therefore we're stuck with
a hardcoded wait time.