Add instructions that Mac devs will hit with the staging environment
While trying to run the staging environment on Mac OS X with Ansible
2, one can run into errors due to changes [0] to how file handles
are used to manage forks in Ansible 2 and the fact that by default
Macs have a very low number of max open files (for me it was 256).

This adds a brief note describing how to increase the limit
without disabling System Integrity Protection which many
guides [1] suggest. I'm happy to remove this once the issue is
resolved or if there is a better guide elsewhere we can link to.
Until then given the number of developers that use Mac OS X, we should
probably include some guidance.

[0] https://docs.ansible.com/ansible/intro_installation.html#control-machine-requirements
[1] https://blog.dekstroza.io/ulimit-shenanigans-on-osx-el-capitan/