Pull in syncronize module from ansible `devel`
The version under ansible <2.3 wasnt compatible with docker so I had
to pull from upstream until we decide to utilize ansible 2.3 in SD. One tweak

I had to make locally was on line 400-404, needed to make sure the --rsh flag
was surrounded by two levels of quotes ("'). I didnt have time to debug this
behavior as it seems it was intentionally removed from upstream in a recent
commit. Might need to debug this when we decide to rip this fork out in the
future.