Added the CI image used in CircleCI for tracking
This is the configuration for the docker image used under CircleCI 2.0.
Specifically, this is the image that circle pulls in and runs all commands from.
It was important this image has basic python tooling, git, and docker. What
happens is that in this environment, the code is cloned down and then a remote
docker environment is setup and all future docker commands are run against that
other docker instance. Basically - docker calls to a remote docker instance
from inside a docker image. Hoorayy!!