Add CI helper script to setup local env
It is sometimes annoying to remember all the environment variables that need to
be added to make local CI testing work. This throws them in a friendly helper
script with prompts. The only friendly default not available is for the VPC ID
but if you already have that set it will try to re-import the previously set one
as a default choice.