Makes haveged config idempotent
We lost the idempotence originally implemented in
4e3ee59df6acecb0dd150f0274464d2edcbdcc23 due to the Ansible v1 -> v2
transition. The escaping syntax for all special characters, including
newlines, has changed, so the awk command in the haveged task list was
inserting a literal '\n' string into the file. Fortunately the second
pass with `lineinfile` cleaned up the error, so the misconfiguration never
affected active systems (even the development VM), but it did break
idempotence, which highlighted the problem.

Adding single-quotes to the parameter values on the lineinfile tasks is
simplmy a best practice, and aids in readability; it doesn't alter the
behavior of that task.