Installs ntp in dev VM via vars
We already have a single point of update for adding package dependencies
to the dev VM, so let's use it. (Unfortunately that same
`development_dependencies` var is reused as build dependencies on the
build VM, but that's a separate problem.) The new approach avoids
unnecessary apt cache refreshs that slow down dev VM provisioning runs.

Updated config tests to ensure that ntp is installed on the dev VM.