Fixes dev reminder about build VM in staging env
The switch from Ansible v1 to v2 broke the helpful message instructing
developers to up the build VM prior to provisioning the staging
machines. The play targeting localhost to inspect the available machines
was simply being skipped. The solution turns out to be extending the
`--limit` argument as specified in the Vagrantfile from "all" to
"all,localhost". Credit to a helpful blogger [0] for sussing out that
the issue was Vagrant-specific.

Also removed a trailing newline on the message itself, simply to clean
it up a bit, since the default output includes a literal '\n'.

Fixes #1706.

[0] http://blog.indrek.io/articles/ansible-skipping-localhost-when-using-vagrant/