testinfra - check source interface reachable
We were missing issues in CI where everything was deployed and testinfra was
passing but the site was not technically up and accessible. This adds basic
functionality to check over tor that the site is up and not throwing errors.
This is only for the source interface, need to make some further tweaks to fix
journalist interface.