Provide time gap between sys reboot and tor tests
Noticed that we werent giving the app-staging box enough time for the tor
service to fully start up and be accessible. Rather than add a sleep, I pushed
the tor results to the bottom of the stack so that by the time they are run -
the tor service is fully started.