Cleans up app functional tests over tor
Tried to unpack the command-building logic a bit to make it more
readable and understandable, since we'll soon be fleshing these tests
out a lot, now that we have a dedicated host for acting as a client in
the functional tests. No alteration of the test behavior, simply
cosmetic changes.

Also inches the ball forward on flake8-compatible config tests, but
we've got a long way to go still.