Simplifies package cleanup in grsecurity role
The Ansible `apt` module supports an "autoremove" parameter, but
also requires another action, such as a package installation or upgrade,
so we can't use that.

The Ansible apt module supports fnmatch for wildcarding, so let's lean
on the module to handle ensuring the non-grsec-patched distro kernel
packages are absent. Allows us to remove the conditional "changed_when"
logic as well, which was poorly structured anyway.

Now's also a good time to collapse the tasks into a single task with a
loop to pass in different args.

Updated associated config tests with smarter logic in the stdout string
comparison, so we don't have to skip the test any longer on app-staging.