Refines apt-mark hold changed status reporting
Since the staging environment installs the locally-built deb packages on
every provisioning run, regardless of whether the code changed (simply
because it's difficult to detect whether the code changed, so it's
better to be safe than sorry), and also performs an apt upgrade every
time, we need to be sure we don't inadvertently switch to the production
packages in the FPF apt repo on upgrade. Thus we mark the locally built
packages as "held", prohibiting upgrades.

While it's not terribly difficult to ascertain whether the "hold" action
resulted in a change, we're constantly wiping "hold" state by forcing
package installation via `dpkg -i <deb>`. Therefore we should check
whether the packages are marked as held prior to installation, and use
that state to compare the changed status.