Fixes /etc/hosts config idempotence
We lost the idempotence initially implemented in
57241076aa4e1ab4467fdb1346970a34fdc26128, due to the upgrade to Ansible
v2. The logic for escaping special characters, such as newlines, has
changed, requiring only one backslash now, rather than the previous two.
Without this change, the config was inappropriately appending a literal
'\n' to the hosts line, which is definitely not what we want.

Fortunately, the misconfiguration never shipped, nor was the system left
in a broken state by the playbook, due to the two-pass task approach.