Creates dedicated playbook for configuring SecureDrop
Allows Admins to enter info interactively, via Ansible's `vars_prompt`
play-level attribute. We cannot use preexisting vars as "default" values
here, which is a bit maddening, but still the streamlined approach of
interactive prompts is quite nice.

Storing the prompted-for vars on disk, so they can be used automatically
in the SecureDrop provisioning playbook run. Using lineinfile rather than
template, to prevent clobbering of additional Admin overrides that were
not specified in the configure playbook logic.