Migrates deprecated vars file in prod playbook
New instances starting with 0.4+ will use the new git-ignored vars file
location `group_vars/all/site-specific`, but instances set up prior to
0.4 will still have production secrets in the `prod-specific.yml` file.
We're deprecating that file because of the git merge conflicts, as well
as the complicated "stash-pull-pop" workflow we demand of Admins.

In order to support existing instances, let's seamlessly copy over the
old prod-specific.yml vars file to the new location, if-and-only-if no
file exists at the new location. Since Ansible reads variables at the
start of every play, the new vars file will be used in subsequent plays,
so Admins do not need to take any special action.