Creates wrapper script for configure playbook
We can't easily create a Python script to perform the validation in real
time (i.e. loop over an input until the value is satisfactory), since
Tails doesn't provide `yaml` or `pyyaml` by default. Therefore let's use
trusty old Bash, and keep it simple: skip prompting for values already
declared.

Admittedly the script is quite minimal, but providing the script target
gives us a point of update for subsequent improvements to the first-run
Admin experience, and we can modify or replace this script as necessary,
without needing to train Admins on a new strategy.