Imports site-specific vars during validation loop
By default Ansible won't reread the vars for localhost, since the
group_vars file was not present on first run. Let's include it in the
context of localhost explicit, but use a `with_first_found` loop to
avoid erroring out on first run if the file had not been created yet.