Adds validation tasks for SMTP relay and SASL domain
Shells out to the `host` command in order to validate these variables.
There's an argument to be made that requriing network connectivity
simply for vars validation will create problems, but in the context of
the deployment, we need network connectivity to provision the SecureDrop
servers anyway. If the `host` lookup fails, that information should be
reported to the Admin prior to running the playbooks, otherwise
debugging the OSSEC email alerts can be tedious and confusing.