Updates gpg2 commands for Tails 3 in validate logic
The whitespace output for displaying full-length fingerprints changes in
Tails 2 versus Tails 3. The newer syntax allows us to remove the sed
pipe for removing whitespace, and simply grep the entire output for the
precise of the line displaying the full fingerprint (with whitespace
prefix). This is safe to do, because the script created a dedicated
keyring strictly for testing, so no other keys other than the one
imported will be present.

Incorporates liberal use of bash traps for cleanup.