Only validate ansible_user when running against servers
The `validate` role previously ran in two different contexts:

  1. In the configure playbook, for validating input.
  2. In the prod playbook, for validating stored vars.

In 1, the target host is localhost running Tails, the Admin Workstation.
In 2, the target host is remote, running Ubuntu, the SecureDrop servers.
Validation of the `ansible_user` var is only relevant to the latter
case, since that value will always be "amnesia" when running against
localhost. So let's only expect it *not* to be amnesia if we're trying
to connect to the Ansible hosts.

It's safer to do the vars validation prior to connecting to any servers,
since the connection will fail if the SSH user value is wrong. To
provide informative error messaging if the `-u <username>` flag is
omitted, we should consider setting `remote_user` at the play level; but
that's difficult to do, given the current ambiguity of the `ssh_users`
var being a list or a string (we should standardize on it being a list).