Don't require root privs for Tails persistence checks
Using become=yes requires prompting for a temporary Tails root
passphrase in order to inspect the contents of persistence.conf. We'll
do that if we have to, but let's try to run the validate logic without
root privileges if at all possible. Checking for the presence of the
`openssh-client` directory adjacent to the `persistence.conf` file is
sufficient indication that the SSH config persistence was enabled when
configuring the volume.