More informative error messages in validate failures
The validate role runs through the list of site-specific variables and
exits when an error is found. Since the vars file containing the
known-bad variable has already been written to disk at this point, the
`./configure` script will *not* prompt for it, creating a loop of
validation failures. Therefore let's instruct the Admin to edit the file
by hand, and point to the specific var name that's causing the trouble.