Adds validation task for pubkey file existence
The compare-fingerprints script assumed that the pubkey files in
question existed, and didn't provide informative error messaging
if they didn't. While we could add file-existence checks to the script,
it's cleaner to keep that script focused on the
fingerprint-matches-pubkey file, and separate check file existence prior
to running the script. That way, we can provide very informative error
messaging in the validate role, using the `msg` parameter for `assert.