Modify admin upgrade behavior and messaging
This commit swaps the weird custom messaging logic previously used to take
advantage of pyhon's logging module instead. Refactored the upgrade logic as
well to always attempt a safe update and report back messaging to user. Hide
a lot of the output behind a debug level which now requires a verbose flag to
view.