Provides helpful messaging in Tails config playbook
Similar to the exit message in the old `install.sh` script, we'll
provide informative output to state that the configuration run is
complete. It's probably not necessary to provide explicit URLs, since we
have desktop icons, but preserving the messages on the web interfaces
just in case it's useful. Omitted the Onion URLs for the SSH addresses,
since we have SSH alias configuration, as well as a dynamic inventory
script incoming.