Fixes Onion URL handling in dynamic inventory script
Should have been included in #1782, but was missed during QA (since
testing in Tails is still difficult, but getting simpler fast). We're
not using a straight tor-generated "hostname" file for reading; for
better or for worse, the SecureDrop playbook formats the hostname files
fetched back from the servers with the `HidServAuth` prefix, so we
actually need the second field in the file, not the first.

Exception handling is in place to catch any IndexErrors if for some
reason the file is not formatted as expected.

Also made a minor stylistic cleanup to move the return statement to the
end of the function, to aid in readability.