Minor misc. fixes to crypto_util
The only change that appears to be of any significance here is removing the
`use_agent=True` keyword argument to the `gnupg.GPG()` initializer. This does
not really change anything however, as python-gnupg now sets this option for us
by default. The ticket linked in the comment this commit removes,
https://github.com/isislovecruft/python-gnupg/issues/96 , was resolved years
ago, and I have confirmed the logging noise remains absent when it is not
explicitly specified.