source: set Source.NUM_WORDS to 7
Instead of keeping the constant in generate_unique_codename
argument (which is never called with anything but the default), create a
NUM_WORDS constant (next to MAX_CODENAME_LEN). The two are related since
MAX_CODENAME_LEN is a safeguard that assumes a given number of words and
a given average word length.

Also modify the test that had the same constant accordingly.