Fix failing test
One of the things accomplished in the last commit, which involved a minor
refactor of the `SecureTemporaryFile` class, was:

> * Ensures :meth:read() cannot be called before :meth:write(). I think this was
> the intended behavior (it's at least the behavior promised by the original
> docstring), so I consider this a bug fixed.

Fixing this bug exposed a problem with one of our tests, documented in the
comment in the diff.

In addition to fixing the test, I also modified and moved it because I believe
it belongs in `test_store.py` rather `test_source.py`; clearly what's actually
being tested is a function of the `store.save_file_submission()` function, and
good unit testing practices say we should test this in isolation of the long
execution chain of `source.submit()`.