Explicity select html5lib as bs4 parser
Previously, bs4 was automatically selecting `html5lib` as a HTML parser, a
Python package which was incidentally provided by a system apt package. This for
one threw a warning because bs4 wants you to explicitly select your own parser,
and also is not how we want to track our dependencies. This commit fixes that by
specifically selecting `html5lib` and tracking it in our testing dependencies
s.t. the version we provide will be used when running the tests.

Fixes #1820.