Instructs pip to upgrade dependencies in Travis
As pointed out by @fowlslegs during review of #1859, we can pass the
`-U` option on the pip install commands within Travis to sidestep the
problem of needing to uninstall certain libs already provided by Travis.
Many of the config tests check for the presence of an exact version, and
in the context of Travis, we lazily conflate the Ansible controller and
the development VM, since the latter is localhost in Travis.

Preserved the `pip freeze -l` commands that @fowlslegs requested be
removed, since they are quite useful in debugging, and I have to add
them every time in a feature branch and wait for a build to run in order
to inspect the build env when problems occur. I'd rather the machine do
that work for me, so I can verify as needed what the pip dependencies
actually are within the Travis environment (which doesn't provide SSH
access to failed builds as CircleCI does).