Replace GzipFile mock with a spy
Thanks to the prior commit's more careful checking of the order of
operations called on a SecureTemporaryFile,
`test_submit_sanitizes_filename` started failing because `gzip.GzipFile`
was being mocked and not actually called. This meant that a
SecureTemporaryFile was created in `save_file_submission` (starting in
the `init` state), but it was never written to (transitioning to the
`write` state) in the unit tests because of the mocked `GzipFile`. Thus,
when `crypto_util.encrypt` tried to `read` from the SecureTemporaryFile,
it failed because it was still in the `init` state, and the improved
checking caught the violation of the intended contract as documented in
SecureTemporaryFile's docstring.

One easy way to fix this is to transform the `GzipFile` mock into a spy
using `wraps`. This allow us to inspect the arguments passed to
`GzipFile`, as needed for our test assertions, but still calls the
underlying `GzipFile` implementation, which fulfills the expected
contract for SecureTemporaryFile and fixes the broken test.