journalist: gracefully handle HOTP format errors
In admin/reset-2fa-hotp, when the token is not properly formatted (odd
number of characters, non hexa character), a python trace is displayed
to the user.

When a known formatting error is detected, a user friendly error message
is displayed instead of the stack trace. Other errors are not caught and
will show a stack trace. Chances are they are not understandable for the
user anyway, even if beautified.