Revert "Instructs pip to upgrade dependencies in Travis"
This reverts commit 4fd5bea894b105ca708c07d436cd40aab83ff478.

The -U flag to pip install is not magic sauce; we still need to
uninstall the pip libs provided by Travis to avoid conflicts. Down the
road we'll work on creating isolated containers for each of these
processes, so they all have their own dependency list independent of
each other.