doc: git describe --exact-match may fail
When not on a tag, it will fail with:

fatal: no tag exactly matches 'dc301206cd38e5f1c30a1687f19d2d0b7b5054cf'

which is not very useful. Changing to git describe alone which will show
the tag if it matches the current commit or something like

0.3.11-2103-gdc301206

otherwise which gives some kind of indication regarding the latest
release.