Ports configure logic to securedrop-admin script
The bash script at `install_files/ansible-base/configure` was already
called via the `securedrop-admin`, providing a single entry point for
Admins to memorize, and for the project documentation to cover. Since
the bash script is not executed directly in the current docs, it's now
been ported to a Python function inside the `securedrop-admin`
management script. Functionality remains the same.