Removes `--user` flag from securedrop-admin
The YAML vars file reading was designed to guard against
misconfigurations, particularly due to the breaking change of forcing
the "ssh_users" var to be a string, rather than a list. We'll still
handle exceptions related to the site-specific vars file, but we no
longer need to read in the username from that file.