Handles Admin username in dynamic inventory
The dynamic inventory script now intelligently reads in the
site-specific YAML vars file and sets the value of the `ssh_users` var
to `ansible_ssh_user` as a host fact. This removes the need for the
Admin to provide the `-u <username>` to all Ansible commands, or the
`--user` flag to securedrop-admin script invocations.