Updates validate role logic for single username
Since we're forcing the `ssh_users` var to contain a single string as
the username, not a list of usernames, we must update the variable
validation logic in order to display informative error messaging about
the breaking change. These checks will run prior to execution of
playbooks against the remote servers, so we should catch all cases of
unsupported vars while operating solely within the Admin Workstation.