Sets Admin username in SSH config under Tails
Allows to remove tons of `<username>` strings throughout the project
documentation. At this point in the installation, the Admin has already
typed the appropriate username and had it recorded in the vars file; we
can always read it from that location, and we do not need to prompt for
it anymore.