Breaks up envsetup function in securedrop-admin
Converts the `envsetup` function, which handles:

  1. Installing apt packages
  2. Creating virtualenv
  3. Installing Python packages via pip

into a wrapper function, which in turn calls discrete functions for the
apt and pip installations. The virtualenv creation is still left in the
function, because it's a brief call.

Also improves exception handling for both new functions. For apt:

During Tails boot, apt packages are sometimes installed automatically,
which results in an apt lock that prohibits interactive apt commands
until those background tasks complete. Mostly the automatic installation
will occur if the Admin configured apt package persistence, which was
configured automatically in SecureDrop 0.3.12 under Tails 2.x, but *not*
in SecureDrop 0.4 under Tails 3.x, since we're switching to a virtualenv
for persistent Admin tooling.