Supports HTTPS in Apache template for Source Interface
Requires new role vars with sane default names for the cert, key, and
chain files. Naming them generically with "SecureDrop" and "Source" so
that any org can use the defaults, as long as the filenames match as
saved on the Admin Workstation.

Passing the vars through the |basename filter in the template
interpolation to ensure that we don't prepending directories when
concatenating the paths for the remote host. The certificates *must*
live in /var/lib/ssl, because that directory is whitelisted in the
apache2 apparmor profile. No other path will work.

Also includes a top-level boolean for enabling or disabling HTTPS
(off by default), which can be overridden during site configuration.
Integration with the `securedrop-admin sdconfig` logic is written.

Conditionally adds 443 to Apache ports.conf for HTTPS

The ports.conf file is included in apache2.conf, and instructs Apache to
listen on certain sockets. By default that's 80 (Source) and 8080
(Journalist), but for HTTPS on the Source Interface, we'll also need a
443 listen declaration.