Clarifies error message for Tails apt installation
Noted during review by @msheiny. The apt lock assumption was a poor one,
since any failure in the apt task whatsoever would allege that the apt
lock caused it. Updated the message to recommend checking the network
connection and trying again, which is consistent with the other common
error messages throughout the securedrop-admin script, and more in line
with the likely culrpit (i.e. networking under Tails).