journalist: use false() instead of == False
The false() function from sqlalchemy should be used for comparision with
booleans:

http://docs.sqlalchemy.org/en/rel_1_0/core/sqlelement.html#sqlalchemy.sql.expression.false

Comparing with "== False" works but triggers a flake8 error:

E712 comparison to False should be 'if cond is False:' or 'if not cond:'

Replacing with "is False" does not work as expected. It is better to
avoid using False entirely in this context: it is too fragile and error
prone.