Adjusts Apache permitted module lists
We were previously blacklisting the "setenvif" module, but the "ssl"
module requires it, so removing "setenvif" from the blacklist.
In general we don't want to enable "ssl" at all times, but it's
difficult to enable conditionally, since Apache configs are read in
during execution of the `apache2_module` Ansible module. If SSL-related
config lines are found in the vhost config, the ssl module *must* be
enabled, otherwise the module activation or deactivation will fail.

Worth considering whether we replace use of the `apache2_module` with
`a2enmod -p` and inspect the output to determine changed status.