Sets Apache listening addresses in app role defaults
The staging environment no longer uses port forwarding for direct access
to the web interfaces. Therefore we can safely standardize on
"127.0.0.1" for binding throughout the Apache config. Let's use a role
default var so we can remove the |default() filters sprinkled around the
templates.

Similarly, we can standardize on localhost for the "allow from"
declaration, since web interface access is only allowed over Tor in
staging. Created a var for that and set to 127.0.0.1, removing all
corresponding |default() filters in the templates.

Updated the testinfra config tests accordingly. The Apache test suite
was still using hardcoded vars in some places, even though we added
dynamic variable reading based on hostname. Updated the relevant tests
to use the YAML vars file, as with the other tests.