i18n: load the locale depending on user preferences
Set the locale of the source or journalist interface as follows, by
order of precedence:

  - l request argument or session['locale']
  - browser suggested locale via Accept-Languages
  - config.DEFAULT_LOCALE
  - 'en_US'

When the l request argument is specified, the locale is saved in the
session. If the l request argument is empty, the locale from the session
is cleared and the Accept-Languages header takes precedence again.

The locale requested with the l argument or via the Accept-Languages
header is matched against the available locales, as found in the
translations directory. The babel.core.negotiate_locale function is used
to find a match and it will, for instance, match:

   fr to fr_FR
   fr-FR to fr_FR
   zh-CN to zh_Hans_CN (script part)
   zh_Hans to zh_Hans

In order to improve the chances to find a match, the Accept-Languages
values are normalized via the core.Locale.parse function which knowns
about aliases, default script for Chinese etc.