Fix broken reply integration tests
html.parser isn't as lenient as html5lib, the HTML parser that
BeautifulSoup was using by default before we started explicitly
specifying html.parser. Switching to a less-lenient parser exposed two
tests that were relying on lenient parsing to succeed:
TestIntegration.test_reply_normal and
TestIntegration.test_unicode_reply_with_ansi_env.

Printing the BeautifulSoup parse tree in the failing tests showed that
it was interpreting `input[name=reply_filename]` as a *child* of
`input[name=csrf_token]`, when it should be a sibling. This was due to
the invalid trailing slash on the `input[name=csrf_token]`. Since the
BeautifulSoup CSS selector was specifically looking for a child (not any
descendant), this unintentional extra layer of nesting caused the
selector, and thus the test, to fail.

This commit also does some minor cleanup of invalid HTML5 in the same
area of code. All tests are now passing with html.parser.