Fixes Ansible groups for prod VMs in Vagrantfile
The changes presented in #1774 updated the prod inventory to use a
dynamic inventory script, but did not include corresponding changes to
the Vagrantfile for the prod VMs. Implemented here, specifically the
"securedrop" group addition, which is used to target both the
Application and the Monitor Servers in production.

In addition to the "securedrop" addition, also amended the "limit"
option in the Ansible config for the prod VMs to include "localhost",
which unintuitively is not considered when handling a limit of "all".
The explicit override enables localhost to be targeted by plays, e.g.
the vars validation logic in the prod playbook.

Closes #1943.