journalist: templates: add id to flag continuation
Add an id to the continuation link of the flag information page of the
journalist interface. This allows test to verify the link is present or
click it without relying on the textual content of the link or its
position in the DOM, both of which are fragile and will break if l10n is
active or the DOM changes.