Quickfix for messy output from apt-get install
-q omits progress indicators, which generally look messy when reprinted.
Note that -q must come *before* the command (e.g. install).

Using -q alone does not solve the problem because apt-get calls out to
dpkg, which gets confused about whether it is printing to a
terminal or not, attempts to use carriage returns to display
incremental/progressive output, and thus causes the output to look
messy. Dpkg::Use-Pty=0 prevents this behavior.

I have tested this on Tails 3.0.1 and confirmed that this change fixes
the messy output from `apt-get install` when it is called from
`securedrop-admin setup`.