Fix tor reboot task for CI
The waiting task was previously pegged to specific situations when an onion
address was being used. Now that this restriction has been removed the handler
was breaking in CI. Fixed this by removing the legacy ansible_ssh_host and
replacing with conditionally injected remote_host_ref. Still default back to the
standard `ansible_host` if we arent in CI.