Removes 'Sandbox' from torrc config test suite
Updates the testinfra test suite to stop checking for presence of the
"Sandbox 1" declaration, which was backed out of the server configs
during QA for 0.4. In addition to simply removing the check, added a new
test to confirm its absence, so we're certain that the option is indeed
not present on systems during testing. Included links to relevant GitHub
issues for discussion, since we plan to re-add the option down the road,
after sufficient testing.