source: not found page content is empty
The #content id is set to hide its content by default, which is good
when used to hide the codename in the lookup.html page. Unfortunately
the same id (#content) is also used in the error.html and notfound.html
templates and hide their content.

Remove the extra <div id="content"> in both error.html and
notfound.html. Also add a functional test to verify the message is
visible when notfound.html is displayed.