source: allow testing of tor2web
Although it is possible to manually install an extension to the tor
browser to add a custom header and set X-tor2web in this way, it makes
the development environment more complicated. Besides, no other
source/journalist feature needs that kind of fixture and the effort does
not seem worth the reward.

Add a global variable TEST_X_TOR2WEB that allows tests to simulate the
presence of tor2web.