tests: functional: do not check text when source logs out
The important class is unique to the notification displayed on
logout. There is no need to verify the text it contains. It is fragile
and will break when the function is used with l10n for another language
or when the text changes in the template. It also breaks depending on
which browser is used because the notification is not displayed unless a
tor browser is used.