tests: functional: wait for app to respond
There is a race between the navigator and the apps. Although rare, it is
possible the app are not responding yet when the first .get() is issued
to the browser. If that happens the test will just stop and fail instead
of waiting for all elements to load.

Poll the app URLs for 30 seconds before returning so there is no race
between the navigator and the apps.