Breaks out info from Tails 3 guide -> 0.4 upgrade guide
The later steps of the Tails 3 upgrade guide were specific to the 0.4
SecureDrop release upgrade, which is confusing and disorganized. Ported
those steps to the dedicated SecureDrop 0.3.x -> 0.4 upgrade guide, and
linked back to the Tails 3 upgrade guide at the top of the document to
indicate clearly that Tails 3 is a prerequisite.