Migrates the ATHS info file Document -> Journalist interface
Since we renamed the Document Interface to the Journalist Interface, we
must ensure that upgrading instances have the proper file on the Tails
workstations for configuring the torrc access. Due consideration was
given to migrating the configs on the Application Server, and this
commit similarly ensures a working setup on the Tails workstations (both
Admin and Journalist). Documentation will require that the
`./securedrop-admin tailsconfig` entrypoint be used for reconfiguring
the Tails workstations after upgrading from Tails 2 to Tails 3.