Migrates deprecated prod-specific.yml vars file
Now using a new, non-versioned-controlled filepath at
`group_vars/all/site-specific`. The new strategy will avoid any
stash/pop conflicts during upgrades of SecureDrop due to customizations
in the old prod-specific.yml vars file. We'll optimistically copy the
over the old vars file, if-and-only-if the new location doesn't exist.

The first-time install docs instruct Admins to run `./securedrop-admin
sdconfig` before running `./securedrop-admin tailsconfig`, and the
latter invokes the logic committed here. Upgrade docs for SecureDrop 0.4
will instruct Admins to run the latter command to ensure a smooth
upgrade.

If-and-only-if the prod-specific.yml file was migrated, we'll need to
import the variables defined in that file, since they were not included
at the start of the play, given that the prod-specific.yml location is
ignored as of 0.4.