Updates tailsconfig command to bypass inventory
The inventory file intentionally throws an exception if the
new site-specific vars file [0] is not found, and instructs
the Admin to run the `sdconfig` command. We do not need
the dynamic inventory script to target the Tails workstation
via localhost, so pass /dev/null in for the inventory file
on that subcommand, same as we do for the `sdconfig` command.

[0] install_files/ansible-base/group_vars/all/site-specific