Handles server sudo password via Ansible
Rather than prompt for the sudo password to be used on remote servers
and pass that to an ansible-playbook subprocess, let's have Ansible
prompt directly for the sudo password. Deferring the password handling
to Ansible saves us from having to sanitize the input, and makes the
behavior far more predictable.

Since we're still prompting for the sudo password regardless of whether
it's actually necessary, added some informative output to the script to
state that the sudo password is not actually necessary at all times.

In the future, we may wish to detect a first-time installation setup by
e.g. testing for the presence of the `app-ssh-aths` file on disk.