Enforces ownership on SecureDrop dotfiles under Tails
The new Ansible logic for setting up the Tails workstation environments
was not explicitly setting file ownership. Since the old SecureDrop
dotfiles directory from 0.3.12 (and Tails 2) will be root-owned, due to
./install.sh being run with sudo, we need to chown that directory to
allow writes to it. Modifying the ownership in a second task, and only
on the files that reside within `/home/amnesia`.