Fixes Application Server backup logic
The `backup` role was using `async` for the creation of the backup
tarball, but failed to wait on the results of that task before
proceeding. Since the subsequent tasks operate on the backup tarball,
the role failed due to an undefined variable, specifically the
`backup_filename` var.

The solution is to preserve the async nature of the task, since it can
indeed take a while to complete, but make sure to poll for results,
and pause execution of the role until the backup has completed.
For reference, see:

  http://docs.ansible.com/ansible/latest/async_status_module.html