Adds backup action to securedrop-admin script
Admins can now run `./securedrop-admin backup` and a tarball will be
created on the Application Server and fetched back to the Admin
Workstation for safe keeping. Subsequent `restore` functionality will
reference this file and use it for recreating the setup if the servers
ever need to be reprovisioned.

In the process, broke out the backup role logic from the primary
`securedrop-prod.yml` playbook, which means we don't have to pass a mess
of CLI flags in order to run a rather simple action.