Adds restore action to securedrop-admin script
Since we now have a `./securedrop-admin backup` option, we must also
provide a similarly convenient `./securedrop-admin restore
<restore_file>` option.

In the process, broke out the "restore" logic from the "backup" role
(which was overloaded to handle both use cases), to be called via a
dedicated playbook. Doing so provides a simple target for the new
securedrop-admin wrapper script.